created response package

+++ pratice/devbook/api/src/models/users.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID        int    `json:"id,omitempty"`
+	ID        uint64 `json:"id,omitempty"`
 	Name      string `json:"name,omitempty"`
 	Nick      string `json:"nick,omitempty"`
 	Email     string `json:"email,omitempty"`
